requirements.txt
asgiref==3.3.2
Django==3.2.1
#gensim==3.8.3
gensim==4.2.0
# numpy==1.19.4
Pillow==8.0.1
pytesseract==0.3.6
pytz==2020.4
scipy==1.5.4
six==1.15.0
smart-open==4.0.1
sqlparse==0.4.1
#environ==1.0
dj-database-url==1.0.0
django-heroku==0.3.1
gunicorn==20.1.0
django-cloudinary-storage==0.3.0
#psycopg2==2.6.2
cloudinary~=1.30.0
djangorestframework~=3.14.0
firebase_admin==6.0.1
scikit-learn==1.1.2
numpy==1.23.5
konlpy==0.6.0
lexrankr==1.0